Completed bar chart view Fixed shuffle and queue song creation Fixed data point

+++ Canary/Model/DataPoint.swift
@@ -1,8 +1,8 @@
 //
-//  DataSet.swift
+//  DataPoint.swift
 //  test
 //
-//  Created by Nifemi Fatoye on 27/08/2019.
+//  Created by Nifemi Fatoye on 21/01/2020.
 //  Copyright © 2019 Nifemi Fatoye. All rights reserved.
 //
 
@@ -10,11 +10,11 @@ import UIKit
 
 struct DataPoint
 {
-    let value : Int
+    let value : Float
     let colour : UIColor
     let title : String
     
-    init(value : Int, colour : UIColor, title : String)
+    init(value : Float, colour : UIColor, title : String)
     {
         self.value = value
         self.colour = colour
